fix : 리액트 쿼리 데이터 에러 시, 404 페이지 이동
