+++ frontend/pages/project/[id]/index.tsx
@@ -19,6 +19,7 @@ import SubBtn from '@/components/button/SubBtn';
 import { errorAlert } from '@/components/alert/Alert';
 import { postStar } from '@/util/api/postStar';
 import ReviewBox from '@/components/common_box/ReviewBox';
+import Custom404 from '@/components/Custom404';
 
 const ViewProject = () => {
   const loggedInUser = useRecoilValue(loggedInUserState);
@@ -104,14 +105,14 @@ const ViewProject = () => {
     }
   };
 
-  if (projectQuery.error) return <Message>잠시 후 다시 시도해주세요.</Message>;
+  if (projectQuery.error) return <Custom404 />;
   if (projectQuery.isLoading || !data || !applyQuery.data)
     return <Message>로딩중입니다.</Message>;
   return (
     <GridBox>
       <Side>
         <AuthorBox
-        totalProject={data.memberInfo.totalProject}
+          totalProject={data.memberInfo.totalProject}
           userId={data.memberInfo.memberId}
           userImg={data.memberInfo.profileImageUrl}
           userName={data.memberInfo.name}
