+++ frontend/pages/users/[id]/index.tsx
@@ -8,14 +8,16 @@ import UserContentBox from '@/components/user/UserContentBox';
 import Message from '@/components/Message';
 import useUser from '@/hooks/react-query/user/useUser';
 import { onChatCreate } from '@/util/chat';
+import Custom404 from '@/components/Custom404';
 
 const UserPage = () => {
   const router = useRouter();
   const id = router.query.id;
   const {
-    getUserById: { data: user, isLoading },
+    getUserById: { data: user, isLoading, isError },
   } = useUser({ id: id ? +id : undefined });
 
+  if (isError) return <Custom404 />;
   if (isLoading) return <Message>로딩중입니다.</Message>;
   return user && !isLoading ? (
     <GridBox>
