+++ frontend/pages/users/me/edit.tsx
@@ -1,12 +1,14 @@
 import styled from 'styled-components';
 import UserEditForm from '@/components/authAction/UserEditForm';
 import useUser from '@/hooks/react-query/user/useUser';
+import Custom404 from '@/components/Custom404';
 
 export default function Edit() {
   const {
-    getMyInfo: { data: user },
+    getMyInfo: { data: user, isError },
   } = useUser({});
 
+  if (isError) return <Custom404 />;
   return <Wrapper>{user && <UserEditForm user={user} />}</Wrapper>;
 }
